Reorganize metadata api and remove unused endpoint.
Move metadata-related api endpoints into api/metadata.
Also move axios wrappers into api/core to avoid circular dependencies.
