+++ app/assets/src/components/common/Header.jsx
@@ -5,7 +5,7 @@ import ToastContainer from "~ui/containers/ToastContainer";
 import BareDropdown from "~ui/controls/dropdowns/BareDropdown";
 import LogoIcon from "~ui/icons/LogoIcon";
 import { openUrl } from "~utils/links";
-import { deleteAsync } from "~/api";
+import { deleteAsync } from "~/api/core";
 import { forbidExtraProps } from "airbnb-prop-types";
 import cs from "./header.scss";
 import cx from "classnames";
