+++ app/assets/src/components/views/samples/MetadataUploadModal/MetadataUploadModal.jsx
@@ -4,7 +4,8 @@ import { keyBy, flow, mapValues, omit } from "lodash/fp";
 import PropTypes from "prop-types";
 import Wizard from "~ui/containers/Wizard";
 import Modal from "~ui/containers/Modal";
-import { uploadMetadataForProject, getSamplesV1 } from "~/api";
+import { getSamplesV1 } from "~/api";
+import { uploadMetadataForProject } from "~/api/metadata";
 import ReviewPage from "./ReviewPage";
 import UploadPage from "./UploadPage";
 import cs from "./metadata_upload_modal.scss";
