+++ app/assets/src/components/views/samples/MetadataUploadModal/UploadPage.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "~/components/utils/propTypes";
 import MetadataUpload from "~/components/common/MetadataUpload";
 import Instructions from "./Instructions";
-import { validateManualMetadataForProject } from "~/api";
+import { validateManualMetadataForProject } from "~/api/metadata";
 import cs from "./metadata_upload_modal.scss";
 
 class UploadPage extends React.Component {
