+++ config/routes.rb
@@ -28,7 +28,6 @@ Rails.application.routes.draw do
     get :raw_results_folder, on: :member
     post :bulk_upload, on: :collection
     post :bulk_upload_with_metadata, on: :collection
-    get :metadata_types_by_host_genome_name, on: :collection
     get :metadata, on: :member
     get :metadata_fields, on: :collection
     get :contig_taxid_list, on: :member
